+++ emcc.py
@@ -1177,6 +1177,10 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
 
     shared.Settings.TARGET_BASENAME = target_basename = unsuffixed_basename(target)
 
+    if shared.Settings.EXTRA_EXPORTED_RUNTIME_METHODS:
+      diagnostics.warning('deprecated', 'EXTRA_EXPORTED_RUNTIME_METHODS is deprecated, please use EXPORTED_RUNTIME_METHODS instead')
+      shared.Settings.EXPORTED_RUNTIME_METHODS += shared.Settings.EXTRA_EXPORTED_RUNTIME_METHODS
+
     final_suffix = get_file_suffix(target)
 
     if has_dash_c or has_dash_S or has_dash_E or '-M' in newargs or '-MM' in newargs:
@@ -1401,6 +1405,11 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       # memory init file is not supported with side modules, must be executable synchronously (for dlopen)
       options.memory_init_file = False
 
+    # If we are including the entire JS library then we know for sure we will, by definition,
+    # require all the reverse dependencies.
+    if shared.Settings.INCLUDE_FULL_LIBRARY:
+      default_setting('REVERSE_DEPS', 'all')
+
     if shared.Settings.MAIN_MODULE or shared.Settings.SIDE_MODULE:
       if shared.Settings.MAIN_MODULE == 1 or shared.Settings.SIDE_MODULE == 1:
         shared.Settings.LINKABLE = 1
@@ -1413,9 +1422,12 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           '$relocateExports',
           '$GOTHandler',
           '$getDylinkMetadata',
+          '__heap_base',
+          '__stack_pointer',
       ]
-      if options.use_closure_compiler:
-        exit_with_error('cannot use closure compiler on shared modules')
+      # This needs to be exported on the Module object too so it's visible
+      # to side modules too.
+      shared.Settings.EXPORTED_FUNCTIONS += ['___heap_base']
       if shared.Settings.MINIMAL_RUNTIME:
         exit_with_error('MINIMAL_RUNTIME is not compatible with relocatable output')
       if shared.Settings.WASM2JS:
@@ -1487,13 +1499,21 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       ]
 
     if shared.Settings.STACK_OVERFLOW_CHECK:
+      # The basic writeStackCookie/checkStackCookie mechanism just needs to know where the end
+      # of the stack is.
       shared.Settings.EXPORTED_FUNCTIONS += ['_emscripten_stack_get_end', '_emscripten_stack_get_free']
+      if shared.Settings.STACK_OVERFLOW_CHECK == 2:
+        # The full checking done by binaryen's `StackCheck` pass also needs to know the base of the
+        # stack.
+        shared.Settings.EXPORTED_FUNCTIONS += ['_emscripten_stack_get_base']
+
+      # We call one of these two functions during startup which caches the stack limits
+      # in wasm globals allowing get_base/get_free to be super fast.
+      # See compiler-rt/stack_limits.S.
       if shared.Settings.RELOCATABLE:
         shared.Settings.EXPORTED_FUNCTIONS += ['_emscripten_stack_set_limits']
       else:
         shared.Settings.EXPORTED_FUNCTIONS += ['_emscripten_stack_init']
-      if shared.Settings.STACK_OVERFLOW_CHECK == 2:
-        shared.Settings.EXPORTED_FUNCTIONS += ['_emscripten_stack_get_base']
 
     if shared.Settings.MODULARIZE:
       if shared.Settings.PROXY_TO_WORKER:
@@ -3065,20 +3085,22 @@ else if (typeof exports === 'object')
 def module_export_name_substitution():
   global final_js
   logger.debug('Private module export name substitution with ' + shared.Settings.EXPORT_NAME)
-  src = open(final_js).read()
+  with open(final_js) as f:
+    src = f.read()
   final_js += '.module_export_name_substitution.js'
   if shared.Settings.MINIMAL_RUNTIME:
     # In MINIMAL_RUNTIME the Module object is always present to provide the .asm.js/.wasm content
     replacement = shared.Settings.EXPORT_NAME
   else:
     replacement = "typeof %(EXPORT_NAME)s !== 'undefined' ? %(EXPORT_NAME)s : {}" % {"EXPORT_NAME": shared.Settings.EXPORT_NAME}
+  src = re.sub(r'{\s*[\'"]?__EMSCRIPTEN_PRIVATE_MODULE_EXPORT_NAME_SUBSTITUTION__[\'"]?:\s*1\s*}', replacement, src)
+  # For Node.js and other shell environments, create an unminified Module object so that
+  # loading external .asm.js file that assigns to Module['asm'] works even when Closure is used.
+  if shared.Settings.MINIMAL_RUNTIME and (shared.Settings.target_environment_may_be('node') or shared.Settings.target_environment_may_be('shell')):
+    src = 'if(typeof Module==="undefined"){var Module={};}\n' + src
   with open(final_js, 'w') as f:
-    src = re.sub(r'{\s*[\'"]?__EMSCRIPTEN_PRIVATE_MODULE_EXPORT_NAME_SUBSTITUTION__[\'"]?:\s*1\s*}', replacement, src)
-    # For Node.js and other shell environments, create an unminified Module object so that
-    # loading external .asm.js file that assigns to Module['asm'] works even when Closure is used.
-    if shared.Settings.MINIMAL_RUNTIME and (shared.Settings.target_environment_may_be('node') or shared.Settings.target_environment_may_be('shell')):
-      src = 'if(typeof Module==="undefined"){var Module={};}\n' + src
     f.write(src)
+  shared.configuration.get_temp_files().note(final_js)
   save_intermediate('module_export_name_substitution')